Added initial support for Core/EWG statuses
Sometimes we confront an issue that might be better addressed by one of the other
working groups such as Core or Evolution.  In such cases, we will now give the
issue a status that reflects the issue has transferred to the other group, and
should track an issue number with that other group.  The issue is deemed in an
'open, but deferred' state until that other group either resolves the issue, or
throws it back to the LWG saying we are the correct forum to reach a resolution.
Following the Portland 2012 meeting, one issue is moving to each status.