Extracted doc title into metadata
Finally extracted the last piece of metadata out of the lists.cpp
source file, and into the xml/lwg-issues.xml document.  Now that
the title for the mailing is part of the metadata, there should be
no need to rebuild the list generation software for each release
but only for adding new features to the software.

Next planned upgrades are to utility the experimental/filesystem
TS, and possibly the fundamentals experimental/string_view.